feat(Lotto): Lotto 유효성 검사 추가
추가 사항
- 로또 번호가 1~45 사이 정수가 아닐 때 예외처리
- 로또 번호가 중복될 때 예외처리
