+++ README.md
@@ -45,6 +45,13 @@
 - [ ] 보너스 번호를 입력받는다.
     - [ ] (공통) 공백을 제거한다.
 
+### 로또
+
+- [x] 1~45 사이의 중복되지 않는 6개의 로또 번호를 결정한다.
+  - [x] 로또 번호가 6개가 아니라면 오류를 출력한다.
+  - [x] 로또 번호가 1~45 사이의 정수가 아니라면 오류를 출력한다.
+  - [x] 중복되는 번호가 있다면 오류를 출력한다.
+
 ### 발행 로또
 
 - [ ] 로또 구입 금액이 올바른지 확인한다.
@@ -52,12 +59,7 @@
     - [ ] 로또 구입 금액이 1,000원 단위로 나뉘어 떨어지지 않으면 오류를 출력한다.
     - [ ] 로또 구입 금액이 10,000원을 넘으면 오류를 출력한다.
 - [ ] 로또를 생성한다.
-- [ ] 로또 목록을 DTO로 반환한다.
-
-### 로또
-
-- [ ] 1~45 사이의 중복되지 않는 6개의 로또 번호를 결정한다.
-    - [ ] 중복되는 번호를 뽑을 시 다시 추첨한다.
+- [ ] 로또 목록을 반환한다.
 
 ### 당첨 로또
